# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Race Analysis Engine",
    "description": "\nAdvanced Formula 1 analytics.\n\nEvery analysis response carries the `engine_version` that produced it. Cached results\nare keyed on that version, so changing a model invalidates its cached values rather\nthan mixing two definitions in one chart.\n\n**On estimates.** Degradation slopes, fuel-corrected lap times and pit loss are modelled\nfrom observed timing, not measured from team telemetry. Fields that carry an estimate\nsay so in their description. In particular the fuel coefficient is a published\n0.030 s/kg prior, not a value fitted from this dataset \u2014 it cannot be fitted from race\ndata, because fuel load is derived from lap number and carries no independent information.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "List Events",
        "description": "Grand prix weekends, newest first.",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one season",
              "title": "Season"
            },
            "description": "Restrict to one season"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Events Api V1 Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "List Sessions",
        "description": "Timed sessions. Every analysis endpoint is keyed by one of these ids.",
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "FP1-3, Q, SQ, S or R",
              "title": "Kind"
            },
            "description": "FP1-3, Q, SQ, S or R"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionOut"
                  },
                  "title": "Response List Sessions Api V1 Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "Get Session",
        "description": "One session.",
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summaries": {
      "get": {
        "tags": [
          "reference"
        ],
        "summary": "List Summaries",
        "description": "Each race in one line, so a list of 44 becomes navigable.\n\nThe mismatch flag is the useful part: a race where the quickest car did not win\nis where the interesting analysis lives.",
        "operationId": "list_summaries_api_v1_summaries_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one season",
              "title": "Season"
            },
            "description": "Restrict to one season"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryOut"
                  },
                  "title": "Response List Summaries Api V1 Summaries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "season_year": {
            "type": "integer",
            "title": "Season Year"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "circuit_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Circuit Key"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "season_year",
          "round",
          "name"
        ],
        "title": "EventOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "FP1-3, Q, SQ, S or R"
          },
          "event_name": {
            "type": "string",
            "title": "Event Name"
          },
          "season_year": {
            "type": "integer",
            "title": "Season Year"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "total_laps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Laps"
          },
          "telemetry_loaded": {
            "type": "boolean",
            "title": "Telemetry Loaded"
          },
          "start_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Utc"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "kind",
          "event_name",
          "season_year",
          "round",
          "telemetry_loaded"
        ],
        "title": "SessionOut"
      },
      "SessionSummaryOut": {
        "properties": {
          "session_id": {
            "type": "integer",
            "title": "Session Id"
          },
          "event_name": {
            "type": "string",
            "title": "Event Name"
          },
          "season_year": {
            "type": "integer",
            "title": "Season Year"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "total_laps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Laps"
          },
          "telemetry_loaded": {
            "type": "boolean",
            "title": "Telemetry Loaded"
          },
          "winner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner"
          },
          "fastest_driver": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fastest Driver",
            "description": "Quickest on corrected pace, which is often not the winner"
          },
          "pace_winner_mismatch": {
            "type": "boolean",
            "title": "Pace Winner Mismatch",
            "description": "True when the quickest car did not win \u2014 the races worth a look",
            "default": false
          },
          "n_retirements": {
            "type": "integer",
            "title": "N Retirements",
            "default": 0
          },
          "optimal_stops": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimal Stops"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline",
            "description": "One sentence describing the race"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "event_name",
          "season_year",
          "round",
          "telemetry_loaded"
        ],
        "title": "SessionSummaryOut",
        "description": "One race reduced to what makes it worth opening.\n\nA list of 44 identically-shaped rows gives a reader no reason to pick one. This\nis the line that does: who won, whether the quickest car actually won, and what\nthe strategy turned on."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}